extracted menu-entries
