+++ package.json
@@ -31,7 +31,7 @@
   },
   "devDependencies": {
     "babel-eslint": "^3.1.23",
-    "brackets-babel-module-formatter": "^1.2.0",
+    "brackets-babel-module-formatter": "^1.3.0",
     "fs-extra": "^0.22.1",
     "gulp": "^3.9.0",
     "gulp-babel": "^5.1.0",
